+++ app/index.tsx
@@ -7,8 +7,8 @@ export default function Index() {
 
   useEffect(() => {
     setTimeout(() => {
-      router.push('/LoginScreen'); // Redirect to login after 3 seconds
-    }, 3000);
+      router.push('/LoginScreen'); // Redirect to login after 2.5 seconds
+    }, 2500);
   }, []);
 
   return <SplashScreen />;
